reverted expandable data table
