+++ src/components/TaxDetails/taxDetailsContent.scss
@@ -1,0 +1,17 @@
+.Layer__TaxDetails {
+  &__TaxDetailsRow {
+    &--operator {
+      &[data-indent-size='0'] {
+        padding-inline-start: 12px;
+      }
+
+      &[data-indent-size='1'] {
+        padding-inline-start: 24px;
+      }
+
+      &[data-indent-size='2'] {
+        padding-inline-start: 48px;
+      }
+    }
+  }
+}
